image-gen-flux-schnell/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

# FLUX requires newer diffusers with FLUX support
diffusers>=0.30.0
transformers>=4.44.0
safetensors==0.4.3
accelerate==0.33.0
Pillow==10.4.0

# FLUX text encoder dependencies
sentencepiece>=0.1.99
protobuf>=4.25.0

httpx==0.27.0
minio==7.2.12
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
